* rankvis: animations for re orderings  * rankvis: build no uses the jar-in-jar provided by eclipse  * rankvis: preparation for freezed columns  * rankvis: clone able columns  * sandbox: support title
git-svn-id: svn://localhost:8080/caleydo_main/trunk@5970 1f7349ae-fd9f-0d40-aeb8-9798e6c0fce3

+++ org.caleydo.core/src/org/caleydo/core/view/opengl/canvas/internal/swt/SWTGLCanvas.java
@@ -58,7 +58,7 @@ final class SWTGLCanvas implements IGLCanvas {
 			public void widgetDisposed(DisposeEvent e) {
 				// problem if the canvas is not the the top level widget, only release(...) will be called -> call it
 				// manually on disposing
-				if (!PlatformUI.getWorkbench().isClosing()) {
+				if (PlatformUI.getWorkbench() != null && !PlatformUI.getWorkbench().isClosing()) {
 					// e.widget.dispose();
 					fireDispose(SWTGLCanvas.this.canvas);
 					SWTGLCanvas.this.canvas.getContext().destroy();
